2402-SIDI/proyecto-final/docker-compose/apps-python/requirements.txt
psycopg2-binary
colorama